refactor: remove models directory
