+++ src/main.ts
@@ -3,5 +3,5 @@ import { SERVICES } from './service.js';
 
 export * from './interfaces/index.js';
 export * from './interfaces/query-params/index.js';
-export * from './models/index.js';
+export * from './error.js';
 export const createRotaCloudClient = createSdkClient(SERVICES);
