--- src/models/index.ts
@@ -1,1 +1,0 @@
-export * from './SDKError.model.js';
